rag_pipeline_ready-made_earlier/Insights-Agent-Flow-Research-main/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
llama_cloud>=1.0
llama-index-core>=0.14.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
streamlit>=1.40.0
pydantic>=2.0
openai>=1.0.0
matplotlib>=3.8.0
plotly>=5.18.0
cryptography>=3.1
pillow>=10.0.0
ragas>=0.2.0
datasets>=2.14.0
python-dotenv>=1.0.0